+++ Runtime/DB/Item DB/WeaponItem.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 using Sirenix.OdinInspector;
 using UnityEngine;
+using UnityEngine.AddressableAssets;
 
 namespace OcDialogue
 {
@@ -16,18 +17,22 @@ namespace OcDialogue
         public float CurrentDurability { get; set; }
         public float Weight => weight;
         public float Equipped { get; set; }
-        public int TotalPower => physicalPower + firePower + icePower + lighteningPower + darkPower;
 
         public int maxDurability = 100;
         public int weight;
-        public int physicalPower;
-        public int firePower;
-        public int icePower;
-        public int lighteningPower;
-        public int darkPower;
-        [Range(0f, 2f)]public float strikeMultiplier = 1f;
-        [Range(0f, 2f)]public float sliceMultiplier = 1f;
-        [Range(0f, 2f)]public float thrustMultiplier = 1f;
+
+        [BoxGroup("Physical Attack"), ShowInInspector, ReadOnly] public float AveragePhysicalAttack => (strikeAttack + sliceAttack + thrustAttack) / 3f;
+        [BoxGroup("Physical Attack")][Range(0f, 200f)]public int strikeAttack = 100;
+        [BoxGroup("Physical Attack")][Range(0f, 200f)]public int sliceAttack = 100;
+        [BoxGroup("Physical Attack")][Range(0f, 200f)]public int thrustAttack = 100;
+        
+        [BoxGroup("Elemental Attack"), ShowInInspector, ReadOnly] public float AverageElementalAttack => (fireAttack + iceAttack + lighteningAttack + darkAttack) / 4f;
+        [BoxGroup("Elemental Attack")] public int fireAttack;
+        [BoxGroup("Elemental Attack")] public int iceAttack;
+        [BoxGroup("Elemental Attack")] public int lighteningAttack;
+        [BoxGroup("Elemental Attack")] public int darkAttack;
+        
+        public AssetReference avatar;
         public override ItemBase GetCopy()
         {
             var copy = CreateInstance<WeaponItem>();
@@ -42,15 +47,16 @@ namespace OcDialogue
             copy.maxDurability = maxDurability;
             copy.weight = weight;
             
-            copy.physicalPower = physicalPower;
-            copy.firePower = firePower;
-            copy.icePower = icePower;
-            copy.lighteningPower = lighteningPower;
-            copy.darkPower = darkPower;
+            copy.fireAttack = fireAttack;
+            copy.iceAttack = iceAttack;
+            copy.lighteningAttack = lighteningAttack;
+            copy.darkAttack = darkAttack;
             
-            copy.strikeMultiplier = strikeMultiplier;
-            copy.sliceMultiplier = sliceMultiplier;
-            copy.thrustMultiplier = thrustMultiplier;
+            copy.strikeAttack = strikeAttack;
+            copy.sliceAttack = sliceAttack;
+            copy.thrustAttack = thrustAttack;
+
+            copy.avatar = avatar;
         }
 #if UNITY_EDITOR
         public override void SetSubTypeFromString(string subtypeName)